pyproject.toml
[project]
name = "ai-repo-manager"
version = "1.0.0"
description = "A Linux desktop application for managing GitHub repositories with semantic search"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.6.0",
    "PyGithub>=2.1.0",
    "chromadb>=0.4.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ai-repo-manager = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
